upgrade from MiniPy 1.2 to 1.3
