+++ sim/py/python_full.html
@@ -53,7 +53,7 @@
 		<!-- <script src="site/js/libs/mousetrap.min.js"></script> -->
 
 		<!-- the magic (ideally) -->
-		<script src="/sim/py/MiniPy/libs/minipy_v1.2.0.js"></script>
+		<script src="/sim/py/MiniPy/libs/minipy_v1.3.0.js"></script>
 
 		<!-- page-specific scripts -->
 		<script src="/sim/py/MiniPy/js/banner.js"></script>
